Slog!  Now as a transparent, reusable writer.
Demo reforged to use this.  Demo now just thinks it's printf'ing to a normal stream; looks great.

Purposefully avoiding any implementations of progress bars, or anything opinionated at all really.  Those might spawn in a contrib/ package tree, but are absolutely not special enough to be in the main bits.  Lambdas and free content ftw!

Signed-off-by: Eric Myhre <hash@exultant.us>